src/middlewares/__tests__/auth.test.ts
import Auth from '@src/helpers/auth';
import authMiddleware from '@src/middlewares/auth';

describe('AuthMiddleware test', () => {
  it('should verify a JWT token and call next function JWT is valid', () => {
    const token = Auth.generateToken({ payload: 'fake-payload' });

    const reqFake = {
      headers: {
        authorization: `Bearer ${token}`,
      },
    };
    const resFake = {};
    const nextFake = jest.fn();
    authMiddleware(reqFake, resFake, nextFake);
    expect(nextFake).toHaveBeenCalled();
  });

  it('should return a 401 status code if there is a problem with token', () => {
    const reqFake = {
      headers: {
        authorization: `Bearer invalid-token`,
      },
    };
    const sendMock = jest.fn();
    const resFake = {
      status: jest.fn(() => ({
        send: sendMock,
      })),
    };
    const nextFake = jest.fn();
    authMiddleware(reqFake, resFake as object, nextFake);
    expect(resFake.status).toHaveBeenCalledWith(401);
    expect(sendMock).toHaveBeenCalledWith({
      code: 401,
      error: 'jwt malformed',
    });
  });

  it('should return a 401 status code if there is not token', () => {
    const reqFake = {
      headers: {
        authorization: `Bearer`,
      },
    };
    const sendMock = jest.fn();
    const resFake = {
      status: jest.fn(() => ({
        send: sendMock,
      })),
    };
    const nextFake = jest.fn();
    authMiddleware(reqFake, resFake as object, nextFake);
    expect(resFake.status).toHaveBeenCalledWith(401);
    expect(sendMock).toHaveBeenCalledWith({
      code: 401,
      error: 'jwt must be provided',
    });
  });

  it('should return a 401 status code if there is not header "authorization"', () => {
    const reqFake = {};
    const sendMock = jest.fn();
    const resFake = {
      status: jest.fn(() => ({
        send: sendMock,
      })),
    };
    const nextFake = jest.fn();
    authMiddleware(reqFake, resFake as object, nextFake);
    expect(resFake.status).toHaveBeenCalledWith(401);
    expect(sendMock).toHaveBeenCalledWith({
      code: 401,
      error: 'jwt must be provided',
    });
  });
});
